rename, refactor [노은수] : 기본 버튼 컴포넌트 이름 변경, 반응형 버튼으로 리팩토링
